+++ src/pages/Toddel/components/CreateToddelDialog.tsx
@@ -1,5 +1,5 @@
 import { zodResolver } from '@hookform/resolvers/zod';
-import { FormFileUpload, FormImageUpload } from '~/components/inputs/Upload';
+// import { FormFileUpload, FormImageUpload } from '~/components/inputs/Upload';
 import { Button } from '~/components/ui/button';
 import { Calendar } from '~/components/ui/calendar';
 import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from '~/components/ui/form';
@@ -138,9 +138,10 @@ const CreateToddelDialog = () => {
               )}
             />
 
-            <FormImageUpload form={form} label='Velg bilde *' name='image' />
+            {/* TODO: Fix file upload */}
+            {/* <FormImageUpload form={form} label='Velg bilde *' name='image' /> */}
 
-            <FormFileUpload accept='application/pdf' form={form} label='Velg PDF *' name='pdf' />
+            {/* <FormFileUpload accept='application/pdf' form={form} label='Velg PDF *' name='pdf' /> */}
 
             <Button className='w-full' disabled={createToddel.isLoading} type='submit'>
               {createToddel.isLoading ? 'Oppretter...' : 'Opprett'}